requirements.txt
google-auth-oauthlib
google-api-python-client
gtts
selenium
webdriver_manager
schedule
python-dotenv
pywhatkit
requests
google-generativeai